client: unify traits across protocol versions
